Split eigen backend implementation
